planet-search/backend/requirements.txt
blinker==1.7.0
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jplephem==2.19
MarkupSafe==2.1.3
numpy==1.26.1
python-dotenv==1.0.0
requests==2.31.0
sgp4==2.22
skyfield==1.46
urllib3==2.0.7
Werkzeug==3.0.1
